% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc.analysis-package.R
\docType{package}
\name{amc.analysis-package}
\alias{amc.analysis}
\alias{amc.analysis-package}
\title{amc.analysis: Synthetic shapefile (AMC) of administrative boundaries with changes over time}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Natali Stigler-Leguizamo \email{Natali.Stigler@gmail.com}

}
\keyword{internal}
