% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_functions.R
\name{amc_compare_threshold}
\alias{amc_compare_threshold}
\title{compare_threshold}
\usage{
amc_compare_threshold(df_overlap, threshold = seq(0, 20, by = 1), plot = TRUE)
}
\arguments{
\item{df_overlap}{Output of function amc_overlap_analysis()}
}
\description{
compare_threshold
}
