% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_functions.R
\name{amc_create_amc}
\alias{amc_create_amc}
\title{Create AMC}
\usage{
amc_create_amc(
  df_overlap,
  shps,
  threshold,
  id_unit_var,
  period_var,
  ref_period
)
}
\arguments{
\item{df_overlap}{Output of function amc_overlap_analysis()}
}
\description{
Create AMC
}
