% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_functions.R
\name{amc_create_crosswalk}
\alias{amc_create_crosswalk}
\title{Make crosswalk}
\usage{
amc_create_crosswalk(
  shps,
  amc_shp,
  id_unit_var,
  period_var,
  threshold,
  output_file = NULL
)
}
\description{
Make crosswalk
}
