% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_functions.R
\name{amc_get_clusters}
\alias{amc_get_clusters}
\title{Get clusters}
\usage{
amc_get_clusters(df_overlap, threshold = NULL)
}
\arguments{
\item{df_overlap}{Output of function amc_overlap_analysis()}
}
\description{
Get clusters
}
