% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_example.R
\name{amc_mean}
\alias{amc_mean}
\title{Example function to calculate mean}
\usage{
amc_mean(x)
}
\arguments{
\item{x}{input}
}
\description{
Example function to calculate mean
}
\examples{
amc_mean(1:5)
}
\references{
See http//:wrong.com
}
