% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_functions.R
\name{amc_overlap_analysis}
\alias{amc_overlap_analysis}
\title{Make overlap analysis}
\usage{
amc_overlap_analysis(
  shps,
  id_unit_var,
  period_var,
  area_threshold = 1,
  output_file = NULL
)
}
\arguments{
\item{shps}{A list of spatial objects (\code{sf}) representing the territorial units for different periods.
Each element of the list corresponds to one period and share the same projected coordinate reference system (CRS).}

\item{id_unit_var}{The identifier variable present in each spatial object that uniquely defines the territorial unit
in a given period.}

\item{period_var}{A numerical variable present in each spatial object that defines the period in which the units are
being represented.}

\item{area_threshold}{A numeric value used to remove very small overlaps. Default is 1.}

\item{output_file}{Optional file path to save the results in \code{.fst} format. If \code{NULL}, the results
are only returned in memory without saving.}
}
\description{
Function overlap_analysis uses the function ovr_get_overlap_pairs to compute the overlap area
between territorial units in consecutive periods.
}
