% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexico_shps.R
\docType{data}
\name{mexico_shps}
\alias{mexico_shps}
\title{Dataset of Mexican municipalities}
\format{
A list of shapefiles.
}
\usage{
data(mexico_shps)
}
\description{
Dataset of Mexican municipalities
}
\keyword{datasets}
