% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_territorial_units}
\alias{plot_territorial_units}
\title{Plot territorial units by period}
\usage{
plot_territorial_units(
  shps,
  id_unit_var,
  period_var,
  show_legend = TRUE,
  show_label = FALSE,
  legend_position = "right"
)
}
\arguments{
\item{shps}{A list of spatial objects (\code{sf}) representing the territorial units for different periods.
Each element of the list corresponds to one period and share the same projected coordinate reference system (CRS).}

\item{id_unit_var}{The identifier variable present in each spatial object that uniquely defines the territorial unit
in a given period.}

\item{period_var}{A numerical variable present in each spatial object that defines the period in which the units are
being represented.
#' @param show_legend Logical. Should the legend be displayed? Default = TRUE.}

\item{legend_position}{Character. Position of the legend if shown. Options: "right", "bottom", "left", "top".
Default = "right".}
}
\description{
Plot territorial units by period
}
